requirements.txt
# DCGAN Image Generation Project Requirements
# Author: RSK World
# Website: https://rskworld.in
# Email: help@rskworld.in
# Phone: +91 93305 39277

torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=10.0.0
jupyter>=1.0.0
notebook>=6.5.0
ipywidgets>=8.0.0
tensorboard>=2.13.0
scipy>=1.10.0
flask>=2.3.0
tqdm>=4.65.0

